+++ src/runner/settings_window.h
@@ -17,6 +17,3 @@ ESettingsWindowNames ESettingsWindowNames_from_string(std::string value);
 
 void open_settings_window(std::optional<std::wstring> settings_window);
 void close_settings_window();
-
-void open_oobe_window();
-void open_scoobe_window();